chore: moved shortcodes into directories
