+++ .eleventy.js
@@ -1,7 +1,7 @@
 const dateFilter = require('./src/filters/date-filter.js');
 const isoDateFilter = require('./src/filters/isodate-filter.js');
-const imageShortcode = require('./src/shortcodes/image.js');
-const socialImageShortcode = require('./src/shortcodes/socialImage.js');
+const imageShortcode = require('./src/shortcodes/image/image.js');
+const socialImageShortcode = require('./src/shortcodes/social-image/socialImage.js');
 const mdIt = require('markdown-it');
 const mdDeflist = require('markdown-it-deflist');
 const mdAttrs = require('markdown-it-attrs');
